+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Zettel
 
 ## [Unreleased]
+
+### Fixed
+- Rendering container fragments
+
+### Added
 - Replace `ListState` with `Value` type for representing content.
+- Added fragments example
+- Added mentions example
 
 ## [0.0.23]